Create README.md
Connect Four is a two-player connection game firstly insert your name, in which players first choose a color and then take turns dropping colored disfrom the top into a seven-column, six-row vertically suspended grid. The pieces fall straight down, occupying the next available space within the column. The objective of the game is to be the first…